test(query-builder): move order-by regression tests into query-builder/order-by/
