+++ test/functional/query-builder/order-by/from-subquery/entity/User.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "../../../../src"
+import { Column, Entity, PrimaryGeneratedColumn } from "../../../../../../src"
 
 @Entity("user")
 export class User {
