+++ test/functional/query-builder/order-by/joined-relation/entity/Photo.ts
@@ -1,4 +1,4 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "../../../../src"
+import { Column, Entity, PrimaryGeneratedColumn } from "../../../../../../src"
 
 @Entity()
 export class Photo {
