+++ test/functional/query-builder/order-by/with-added-select/entity/Category.ts
@@ -1,4 +1,4 @@
-import { Entity, PrimaryGeneratedColumn } from "../../../../src/index"
+import { Entity, PrimaryGeneratedColumn } from "../../../../../../src/index"
 
 @Entity()
 export class Category {
