+++ test/functional/query-builder/order-by/with-added-select/entity/Post.ts
@@ -1,0 +1,24 @@
+import {
+    Column,
+    Entity,
+    PrimaryGeneratedColumn,
+} from "../../../../../../src/index"
+import { Category } from "./Category"
+import { ManyToMany } from "../../../../../../src/decorator/relations/ManyToMany"
+import { JoinTable } from "../../../../../../src/decorator/relations/JoinTable"
+
+@Entity()
+export class Post {
+    @PrimaryGeneratedColumn()
+    id: number
+
+    @Column()
+    name: string
+
+    @Column()
+    count: number
+
+    @ManyToMany(() => Category)
+    @JoinTable()
+    categories: Category[]
+}
