--- test/github-issues/1259/entity/Post.ts
@@ -1,20 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "../../../../src/index"
-import { Category } from "./Category"
-import { ManyToMany } from "../../../../src/decorator/relations/ManyToMany"
-import { JoinTable } from "../../../../src/decorator/relations/JoinTable"
-
-@Entity()
-export class Post {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-
-    @Column()
-    count: number
-
-    @ManyToMany(() => Category)
-    @JoinTable()
-    categories: Category[]
-}
