--- test/github-issues/9960/entity/ExampleEntity.ts
@@ -1,12 +1,0 @@
-import { Entity } from "../../../../src/decorator/entity/Entity"
-import { Column } from "../../../../src/decorator/columns/Column"
-import { PrimaryGeneratedColumn } from "../../../../src/decorator/columns/PrimaryGeneratedColumn"
-
-@Entity()
-export class ExampleEntity {
-    @PrimaryGeneratedColumn()
-    id: number
-
-    @Column()
-    name: string
-}
